pyproject.toml
[project]
name = "proposal-ppt"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "streamlit>=1.32.0",
    "python-pptx>=0.6.23",
    "slack-sdk>=3.27.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "lxml>=5.0.0",
    "pypdf2>=3.0.0",
    "docx2txt>=0.8",
    "openpyxl>=3.1.0",
    "playwright>=1.40.0",
]
